Create pic2pix.py
This script will extract people/drawings out of images, filter them, making it usable as sprites in 2d game engines. The images should have distinct colors; the part to extract should not be close in color to the background to be removed; it can be fixed by increasing TOLERANCE though. 
For monocolor pictures, like pencil doodles on paper, change the name of the image to something with pencil, this triggers a different setting to work on drawings.